Fixed size_t conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@64200 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objmgr/impl/handle_range_map.hpp
@@ -63,8 +63,8 @@ public:
     void AddSegmentIds(const TIds& ids);
     void AddSegmentIds(const TIds2& ids);
 
-    size_t GetSegmentCount(void) const {
-        return m_SegSet.size();
+    int GetSegmentCount(void) const {
+        return int(m_SegSet.size());
     }
     int FindSeg(const CSeq_id_Handle& h) const;
     bool GetMinusStrand(int seg) const;
